Implemented InfluxManager. File attributes. Model adaptions.

+++ Sources/EnergyPackage/EnergyPackage.swift
@@ -1,3 +1,10 @@
+//
+//  Common.swift
+//
+//
+//  Created by Dominik Liehr on 15.10.23.
+//
+
 public final class Common {
-	public static var pv1wattMax: Int = 0
+	public static var pv1wattMax: Int64 = 0
 }
